меню, доработки по проверкам и др.

+++ Back/data_source_class_queries.js
@@ -57,17 +57,58 @@ const createDataSourceClass = (request, response) => {
 }
 
 const deleteDataSourceClass = (request, response) => {
-  const id = parseInt(request.params.id)
-  pool.query('DELETE FROM nucl.data_source_class WHERE id = $1', [id], (error, results) => {
-    if (error) {
-      response.status(400).send(`Связь с источником данных не удалена: ${error.message}`);
+  const id = parseInt(request.params.id);
+  pool.query(' SELECT data_source_id, rec_id, table_name FROM nucl.data_source_class where id = $1 LIMIT 1', [id], (err, res) => 
+  {
+    if (res.rowCount>0)
+    {
+      const { data_source_id, rec_id, table_name } = res.rows[0];
+      console.log( 'table_name, rec_id, data_source_id='+table_name + ' ' + rec_id + ' ' + data_source_id  );
+      pool.query('SELECT (SELECT COUNT(id) FROM nucl.value_int_dose WHERE data_source_id = $1 and '+table_name+'_id = $2) AS int_dose_count, '+
+      '(SELECT COUNT(id) FROM nucl.value_ratio_git WHERE data_source_id = $1 and '+table_name+'_id = $2) AS ratio_git_count, '+
+      '(SELECT COUNT(id) FROM nucl.value_ext_dose WHERE data_source_id = $1 and '+table_name+'_id = $2) AS ext_dose_count '+
+      'FROM nucl.value_int_dose LIMIT 1', [data_source_id, rec_id], (err, res) => 
+      {
+        if (res.rowCount>0)
+        {
+          const { int_dose_count, ratio_git_count, ext_dose_count } = res.rows[0];
+          console.log( 'tint_dose_count, ratio_git_count, ext_dose_count=', int_dose_count, ratio_git_count, ext_dose_count );
+          if  (int_dose_count==0 && ratio_git_count==0 && ext_dose_count==0) 
+          {
+            pool.query('DELETE FROM nucl.data_source_class WHERE id = $1', [id], (error, results) => {
+              if (error) {
+                response.status(400).send(`Связь с источником данных не удалена: ${error.message}`);
+              }
+              else {
+                if (results.rowCount == 1)
+                  response.status(200).send(`Связь с источником данных ${id} удалена: cтрок удалено: ${results.rowCount} `);
+                if (results.rowCount == 0)
+                  response.status(400).send(`Запись с кодом ${id} не найдена `)
+              }
+            })          
+          }
+          else
+          {
+            if (int_dose_count!=0) {
+              response.status(400).send(`Запись с кодом ${id} не удалена: имеются ${int_dose_count} связанные записи в таблице "Значения дозовых коэффициентов для внутреннего облучения" `);
+            } else if (ratio_git_count!=0) {
+              response.status(400).send(`Запись с кодом ${id} не удалена: имеются ${ratio_git_count} связанные записи в таблице "Значения коэффициента поглощения в желудочно-кишечном тракте (ЖКТ)" `);
+            } else if (ext_dose_count!=0) {
+              response.status(400).send(`Запись с кодом ${id} не удалена: имеются ${ext_dose_count} связанные записи в таблице "Значения дозовых коэффициентов для внешнего облучения" `);
+            }
+          } 
+        }
+        else
+        {
+          response.status(400).send(`Запись с кодом ${id}: ошибка при проверка связанных таблиц `);
+        }
+      })
     }
-    else {
-      if (results.rowCount == 1)
-        response.status(200).send(`Связь с источником данных ${id} удалена: cтрок удалено: ${results.rowCount} `);
-      if (results.rowCount == 0)
-        response.status(400).send(`Запись с кодом ${id} не найдена `)
+    else
+    {
+      response.status(400).send(`Запись с кодом ${id} не найдена `)      
     }
+
   })
 }
